** ReferenceSpace:     -- Add Point Mapping from ABC space to XYZ space     -- Handle Elements with no Face
** LineBasis:
    -- Works again

** Jacobian and GroupOfJacobian:
    -- Computes Good Jacobian from Basis ReferenceSpace

** Formulation:
    -- Addapted for good Jacobian

** TermProjection:
    -- Addapted to compute function to project at good point (see ReferenceSpace ABCtoXYZ mapping)

** Laplace, Poisson, SteadyWave:
    -- Passed with all basis

** WARNING:
    -- Tet Basis NOT Retested (but seems to work)
    -- FunctionSpace::getDof doesn't work if wrong element type (because of basis implication)
    -- SystemAbstract::fixCoef doesn't work anymore (see above)
    -- Problem with cureved Element (getTypeForMSH gives HO element for FunctionSpace::getDof)

+++ FunctionSpace/LineEdgeBasis.h
@@ -24,7 +24,7 @@ class LineEdgeBasis: public BasisHierarchical1Form{
   //! @param order The order of the Basis
   //!
   //! Returns a new Edge-Basis for Lines of the given order
-  LineEdgeBasis(unsigned int order);
+  LineEdgeBasis(size_t order);
 
   //! Deletes this Basis
   //!
